pyproject.toml
[project]
name = "crypto-daily-analysis"
version = "1.0.0"
description = "AI-powered cryptocurrency daily trend analysis system"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["cryptocurrency", "bitcoin", "ai", "analysis", "trading"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "schedule>=1.2.1",
    "openai>=1.12.0",
    "pycoingecko>=3.1.0",
    "feedparser>=6.0.11",
    "rich>=13.7.0",
    "python-dateutil>=2.8.2",
]

[project.urls]
Homepage = "https://github.com/yourusername/crypto-daily-analysis"
Repository = "https://github.com/yourusername/crypto-daily-analysis"
Issues = "https://github.com/yourusername/crypto-daily-analysis/issues"

[tool.ruff]
# 目标 Python 版本
target-version = "py38"

# 每行最大字符数
line-length = 100

# 排除的目录
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "reports",
    "*.egg-info",
]

[tool.ruff.lint]
# 启用的规则集
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]

# 忽略的规则
ignore = [
    "E501",  # 行太长（由 formatter 处理）
    "B008",  # 函数调用中的默认参数
    "C901",  # 函数太复杂
]

[tool.ruff.format]
# 使用双引号
quote-style = "double"

# 缩进使用空格
indent-style = "space"

[tool.bandit]
# Bandit 安全检查配置
exclude_dirs = [
    "reports",
    "__pycache__",
    ".venv",
    "venv",
]

# 跳过的测试
skips = [
    "B101",  # assert_used
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"
